Address PR Review comments
