+++ docs/docs/others/telemetry.mdx
@@ -12,5 +12,5 @@ We do not collect any email, first name, last name, phone number, date of birth,
 # Opting-out of telemetry
 Opting out is simple. To do this, edit your .env on the server side and include the following:
 ```
-TELEMETRY_ENABLED=false
+IS_TELEMETRY_ENABLED=false
 ```
