+++ front/src/modules/events/hooks/useIsTelemetryEnabled.ts
@@ -1,4 +1,4 @@
 export function useIsTelemetryEnabled() {
   // TODO: replace by clientConfig
-  return process.env.TELEMETRY_ENABLED !== 'false';
+  return process.env.IS_TELEMETRY_ENABLED !== 'false';
 }
