+++ front/src/modules/events/hooks/useTrackEvent.ts
@@ -1,0 +1,7 @@
+import { useEventTracker } from './useEventTracker';
+
+export function useTrackEvent(eventType: string, eventData: any) {
+  const eventTracker = useEventTracker();
+
+  return eventTracker(eventType, eventData);
+}
