+++ front/src/modules/events/hooks/useTrackPageView.ts
@@ -1,27 +1,13 @@
 import { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 
-import { useCreateEventMutation } from '~/generated/graphql';
-
-import { useIsTelemetryEnabled } from './useIsTelemetryEnabled';
+import { useEventTracker } from './useEventTracker';
 
 export function useTrackPageView() {
   const location = useLocation();
-  const telemetryEnabled = useIsTelemetryEnabled();
-
-  const [createEventMutation] = useCreateEventMutation();
+  const eventTracker = useEventTracker();
 
   useEffect(() => {
-    if (!telemetryEnabled) {
-      return;
-    }
-    createEventMutation({
-      variables: {
-        type: 'pageview',
-        data: JSON.stringify({
-          location: location,
-        }),
-      },
-    });
-  }, [location, createEventMutation, telemetryEnabled]);
+    eventTracker('pageview', { location });
+  }, [location, eventTracker]);
 }
